fixed compliance tests
